@@ -9,12 +9,12 @@ buildscript {
         }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath 'com.android.tools.build:gradle:3.5.0'
         classpath 'com.github.dcendents:android-maven-gradle-plugin:1.5'
         classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.0'
         classpath 'org.greenrobot:greendao-gradle-plugin:3.2.2' // add plugin
       //  classpath 'com.bugsnag:bugsnag-android-gradle-plugin:3.4.1'
-        classpath 'com.tencent.bugly:symtabfileuploader:2.1.2'
+        classpath 'com.tencent.bugly:symtabfileuploader:2.2.1'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -39,23 +39,16 @@ task clean(type: Delete) {
 // Define versions in a single place
 ext {
     // Sdk and tools
-    minSdkVersion = 10
-    targetSdkVersion = 25
-    compileSdkVersion = 25
-    buildToolsVersion = '26.0.2'
+    minSdkVersion = 19
+    targetSdkVersion = 28
+    compileSdkVersion = 28
+    buildToolsVersion = '28.0.3'
 
     // App dependencies
-    supportLibraryVersion = '26.+'
-    guavaVersion = '18.0'
-    junitVersion = '4.12'
-    mockitoVersion = '1.10.19'
-    powerMockito = '1.6.2'
-    hamcrestVersion = '1.3'
-    runnerVersion = '0.5'
-    rulesVersion = '0.5'
-    espressoVersion = '2.2.2'
+    supportLibraryVersion = '28.0.0'
     exoPlayerVersion = "2.6.0"
-    butterKnifeVersion = '8.8.1'
-    daggerVersion = '2.16'
+    butterKnifeVersion = '9.0.0'
+    daggerVersion = '2.21'
+    retrofitVersion='2.5.0'
     markwonVersion = '2.0.1'
 }
