@@ -7,6 +7,7 @@ import com.danikula.videocache.HttpProxyCacheServer;
 import com.danikula.videocache.headers.HeaderInjector;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.u9porn.constants.Constants;
 import com.u9porn.cookie.AppCookieManager;
 import com.u9porn.cookie.CookieManager;
 import com.u9porn.data.AppDataManager;
@@ -14,7 +15,6 @@ import com.u9porn.data.DataManager;
 import com.u9porn.data.cache.CacheProviders;
 import com.u9porn.data.db.AppDbHelper;
 import com.u9porn.data.db.DbHelper;
-import com.u9porn.data.model.User;
 import com.u9porn.data.network.ApiHelper;
 import com.u9porn.data.network.AppApiHelper;
 import com.u9porn.data.prefs.AppPreferencesHelper;
@@ -26,13 +26,8 @@ import com.u9porn.utils.AddressHelper;
 import com.u9porn.utils.AppCacheUtils;
 import com.u9porn.utils.MyHeaderInjector;
 import com.u9porn.utils.VideoCacheFileNameGenerator;
-import com.u9porn.constants.Constants;
 
 import java.io.File;
-import java.net.Proxy;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
 
 import javax.inject.Singleton;
 
@@ -48,6 +43,8 @@ import io.victoralbertos.jolyglot.GsonSpeaker;
  */
 @Module
 public abstract class ApplicationModule {
+
+    private static final String TAG = ApplicationModule.class.getSimpleName();
 
     @Binds
     @ApplicationContext
@@ -132,4 +129,28 @@ public abstract class ApplicationModule {
     static CookieManager providesCookieManager(AppCookieManager appCookieManager) {
         return appCookieManager;
     }
+
+//    @SuppressLint("SetJavaScriptEnabled")
+//    @Provides
+//    @Singleton
+//    static WebView providesWebView(@ApplicationContext Context context){
+//        Logger.t(TAG).d("初始化");
+//        WebView mWebView = new WebView(context);
+//
+//        WebSettings mWebSettings = mWebView.getSettings();
+//
+//        //启用JavaScript。
+//        mWebSettings.setJavaScriptEnabled(true);
+//        mWebSettings.setUseWideViewPort(true);
+//        mWebSettings.setJavaScriptCanOpenWindowsAutomatically(true);
+//
+//        mWebView.loadUrl("file:///android_asset/web/index.html"); //js文件路径
+//        mWebView.setWebViewClient(new WebViewClient() {
+//            @Override
+//            public void onPageFinished(WebView view, String url) {
+//                Logger.t(TAG).d("加载完成..:" + url);
+//            }
+//        });
+//        return mWebView;
+//    }
 }
