@@ -210,29 +210,32 @@ public class ParseV9PronVideo {
             videoResult.setId(VideoResult.OUT_OF_WATCH_TIMES);
             return videoResult;
         }
+        Document doc = Jsoup.parse(html);
+        String videoUrl = doc.select("video").first().select("source").first().attr("src");
+        if(!videoUrl.startsWith("http")){
+            final String reg = "document.write\\(strencode\\(\"(.+)\",\"(.+)\",.+\\)\\);";
+            Pattern p = Pattern.compile(reg);
+            Matcher m = p.matcher(html);
+            String param1 = "", param2 = "";
+            if(m.find()){
+                param1 = m.group(1);
+                param2 = m.group(2);
+            }
+            param1 = new String(Base64.decode(param1.getBytes(),Base64.DEFAULT));
+            String source_str = "";
+            for (int i = 0,k=0; i<param1.length(); i++) {
+                k = i % param2.length();
+                source_str += ""+(char)(param1.codePointAt(i) ^ param2.codePointAt(k));
+            }
+            Logger.t(TAG).d("视频source1：" + source_str);
+            source_str = new String(Base64.decode(source_str.getBytes(),Base64.DEFAULT));
+            Logger.t(TAG).d("视频source2：" + source_str);
 
 
-        final String reg = "document.write\\(strencode\\(\"(.+)\",\"(.+)\",.+\\)\\);";
-        Pattern p = Pattern.compile(reg);
-        Matcher m = p.matcher(html);
-        String param1 = "", param2 = "";
-        if(m.find()){
-            param1 = m.group(1);
-            param2 = m.group(2);
+            Document source = Jsoup.parse(source_str);
+            videoUrl = source.select("source").first().attr("src");
         }
-        param1 = new String(Base64.decode(param1.getBytes(),Base64.DEFAULT));
-        String source_str = "";
-        for (int i = 0,k=0; i<param1.length(); i++) {
-            k = i % param2.length();
-            source_str += ""+(char)(param1.codePointAt(i) ^ param2.codePointAt(k));
-        }
-        Logger.t(TAG).d("视频source1：" + source_str);
-        source_str = new String(Base64.decode(source_str.getBytes(),Base64.DEFAULT));
-        Logger.t(TAG).d("视频source2：" + source_str);
 
-//        String videoUrl = doc.select("video").first().select("source").first().attr("src");
-        Document source = Jsoup.parse(source_str);
-        String videoUrl = source.select("source").first().attr("src");
         videoResult.setVideoUrl(videoUrl);
         Logger.t(TAG).d("视频链接：" + videoUrl);
 
@@ -243,7 +246,7 @@ public class ParseV9PronVideo {
         Logger.t(TAG).d("视频Id：" + videoId);
 
         //这里解析的作者id已经变了，非纯数字了
-        Document doc = Jsoup.parse(html);
+
         String ownerUrl = doc.select("a[href*=UID]").first().attr("href");
         String ownerId = ownerUrl.substring(ownerUrl.indexOf("=") + 1, ownerUrl.length());
         videoResult.setOwnerId(ownerId);
