@@ -2,15 +2,12 @@ package com.u9porn.ui.images.viewimage;
 
 import android.annotation.SuppressLint;
 import android.annotation.TargetApi;
-import android.content.DialogInterface;
-import android.content.Intent;
 import android.media.MediaScannerConnection;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.v4.content.FileProvider;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AlertDialog;
 import android.view.View;
@@ -24,12 +21,12 @@ import com.qmuiteam.qmui.widget.dialog.QMUIDialog;
 import com.sdsmdg.tastytoast.TastyToast;
 import com.u9porn.R;
 import com.u9porn.adapter.PictureAdapter;
+import com.u9porn.constants.Keys;
 import com.u9porn.data.model.Mm99;
 import com.u9porn.ui.MvpActivity;
 import com.u9porn.utils.DialogUtils;
 import com.u9porn.utils.GlideApp;
 import com.u9porn.utils.SDCardUtils;
-import com.u9porn.constants.Keys;
 import com.u9porn.widget.ProblematicViewPager;
 
 import java.io.File;
@@ -117,6 +114,9 @@ public class PictureViewerActivity extends MvpActivity<PictureViewerView, Pictur
 
             @Override
             public void onImageLongClick(View view, int position) {
+                if (imageList == null) {
+                    return;
+                }
                 showSavePictureDialog(imageList.get(position));
             }
         });
@@ -151,32 +151,29 @@ public class PictureViewerActivity extends MvpActivity<PictureViewerView, Pictur
 
     private void showSavePictureDialog(final String imageUrl) {
         QMUIDialog.MenuDialogBuilder builder = new QMUIDialog.MenuDialogBuilder(this);
-        builder.addItem("保存图片", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                GlideApp.with(PictureViewerActivity.this).downloadOnly().load(Uri.parse(imageUrl)).into(new SimpleTarget<File>() {
-                    @Override
-                    public void onResourceReady(@NonNull File resource, @Nullable Transition<? super File> transition) {
-                        File filePath = new File(SDCardUtils.DOWNLOAD_IMAGE_PATH);
-                        if (!filePath.exists()) {
-                            if (!filePath.mkdirs()) {
-                                showMessage("创建文件夹失败了", TastyToast.ERROR);
-                                return;
-                            }
-                        }
-                        File file = new File(filePath, UUID.randomUUID().toString() + ".jpg");
-                        try {
-                            FileUtils.copyFile(resource, file);
-                            showMessage("保存图片成功了", TastyToast.SUCCESS);
-                            notifySystemGallery(file);
-                        } catch (IOException e) {
-                            e.printStackTrace();
-                            showMessage("保存图片失败了", TastyToast.ERROR);
+        builder.addItem("保存图片", (dialog, which) -> {
+            GlideApp.with(PictureViewerActivity.this).downloadOnly().load(Uri.parse(imageUrl)).into(new SimpleTarget<File>() {
+                @Override
+                public void onResourceReady(@NonNull File resource, @Nullable Transition<? super File> transition) {
+                    File filePath = new File(SDCardUtils.DOWNLOAD_IMAGE_PATH);
+                    if (!filePath.exists()) {
+                        if (!filePath.mkdirs()) {
+                            showMessage("创建文件夹失败了", TastyToast.ERROR);
+                            return;
                         }
                     }
-                });
-                dialog.dismiss();
-            }
+                    File file = new File(filePath, UUID.randomUUID().toString() + ".jpg");
+                    try {
+                        FileUtils.copyFile(resource, file);
+                        showMessage("保存图片成功了", TastyToast.SUCCESS);
+                        notifySystemGallery(file);
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                        showMessage("保存图片失败了", TastyToast.ERROR);
+                    }
+                }
+            });
+            dialog.dismiss();
         });
         builder.show();
     }
@@ -189,7 +186,7 @@ public class PictureViewerActivity extends MvpActivity<PictureViewerView, Pictur
 
     private void notifySystemGallery(File file) {
 
-        MediaScannerConnection.scanFile(this,new String[]{file.getAbsolutePath()},new String[]{"image/jpeg"},null);
+        MediaScannerConnection.scanFile(this, new String[]{file.getAbsolutePath()}, new String[]{"image/jpeg"}, null);
     }
 
     protected void goFullScreen() {
@@ -206,22 +203,18 @@ public class PictureViewerActivity extends MvpActivity<PictureViewerView, Pictur
 
     private void setUiFlags(boolean fullscreen) {
         View decorView = getWindow().getDecorView();
-        if (decorView != null) {
-            decorView.setSystemUiVisibility(fullscreen ? getFullscreenUiFlags() : View.SYSTEM_UI_FLAG_VISIBLE);
-        }
+        decorView.setSystemUiVisibility(fullscreen ? getFullscreenUiFlags() : View.SYSTEM_UI_FLAG_VISIBLE);
     }
 
     @TargetApi(Build.VERSION_CODES.ICE_CREAM_SANDWICH)
     private int getFullscreenUiFlags() {
         int flags = View.SYSTEM_UI_FLAG_LOW_PROFILE | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION;
 
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            flags |= View.SYSTEM_UI_FLAG_LAYOUT_STABLE
-                    | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-                    | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
-                    | View.SYSTEM_UI_FLAG_FULLSCREEN
-                    | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION;
-        }
+        flags |= View.SYSTEM_UI_FLAG_LAYOUT_STABLE
+                | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
+                | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
+                | View.SYSTEM_UI_FLAG_FULLSCREEN
+                | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION;
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
             flags |= View.SYSTEM_UI_FLAG_IMMERSIVE;
         }
