@@ -1,10 +1,11 @@
 package com.u9porn.ui.porn9video.play;
 
 import android.arch.lifecycle.Lifecycle;
-import android.support.annotation.NonNull;
 import android.text.TextUtils;
+import android.webkit.WebView;
 
 import com.hannesdorfmann.mosby3.mvp.MvpBasePresenter;
+import com.orhanobut.logger.Logger;
 import com.sdsmdg.tastytoast.TastyToast;
 import com.trello.rxlifecycle2.LifecycleProvider;
 import com.u9porn.data.DataManager;
@@ -18,17 +19,19 @@ import com.u9porn.rxjava.RxSchedulersHelper;
 import com.u9porn.ui.download.DownloadPresenter;
 import com.u9porn.ui.porn9video.favorite.FavoritePresenter;
 
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+
 import java.util.Date;
 
 import javax.inject.Inject;
 
 import io.reactivex.disposables.Disposable;
-import io.reactivex.functions.Function;
 
 /**
  * @author flymegoc
  * @date 2017/11/15
- * @describe
+ * @describe play
  */
 public class PlayVideoPresenter extends MvpBasePresenter<PlayVideoView> implements IPlay {
 
@@ -39,8 +42,9 @@ public class PlayVideoPresenter extends MvpBasePresenter<PlayVideoView> implemen
 
     private LifecycleProvider<Lifecycle.Event> provider;
 
-    private int start = 1;
     private DataManager dataManager;
+
+//    private final WebView webView;
 
     @Inject
     public PlayVideoPresenter(FavoritePresenter favoritePresenter, DownloadPresenter downloadPresenter, LifecycleProvider<Lifecycle.Event> provider, DataManager dataManager) {
@@ -54,57 +58,60 @@ public class PlayVideoPresenter extends MvpBasePresenter<PlayVideoView> implemen
     public void loadVideoUrl(final V9PornItem v9PornItem) {
         String viewKey = v9PornItem.getViewKey();
         dataManager.loadPorn9VideoUrl(viewKey)
-                .map(new Function<VideoResult, VideoResult>() {
-                    @Override
-                    public VideoResult apply(VideoResult videoResult) throws VideoException {
-                        if (TextUtils.isEmpty(videoResult.getVideoUrl())) {
-                            if (VideoResult.OUT_OF_WATCH_TIMES.equals(videoResult.getId())) {
-                                //尝试强行重置，并上报异常
-                                dataManager.resetPorn91VideoWatchTime(true);
-                               // Bugsnag.notify(new Throwable(TAG + "Ten videos each day address: " + dataManager.getPorn9VideoAddress()), Severity.WARNING);
-                                throw new VideoException("观看次数达到上限了,请更换地址或者代理服务器！");
-                            } else {
-                                throw new VideoException("解析视频链接失败了");
-                            }
+                .map(videoResult -> {
+                    if (TextUtils.isEmpty(videoResult.getVideoUrl())) {
+                        if (VideoResult.OUT_OF_WATCH_TIMES.equals(videoResult.getId())) {
+                            //尝试强行重置，并上报异常
+                            dataManager.resetPorn91VideoWatchTime(true);
+                            // Bugsnag.notify(new Throwable(TAG + "Ten videos each day address: " + dataManager.getPorn9VideoAddress()), Severity.WARNING);
+                            throw new VideoException("观看次数达到上限了,请更换地址或者代理服务器！");
+                        } else {
+                            throw new VideoException("解析视频链接失败了");
                         }
-                        return videoResult;
                     }
+                    return videoResult;
                 })
                 .retryWhen(new RetryWhenProcess(RetryWhenProcess.PROCESS_TIME))
-                .compose(RxSchedulersHelper.<VideoResult>ioMainThread())
-                .compose(provider.<VideoResult>bindUntilEvent(Lifecycle.Event.ON_DESTROY))
+                .compose(RxSchedulersHelper.ioMainThread())
+                .compose(provider.bindUntilEvent(Lifecycle.Event.ON_DESTROY))
                 .subscribe(new CallBackWrapper<VideoResult>() {
                     @Override
                     public void onBegin(Disposable d) {
-                        ifViewAttached(new ViewAction<PlayVideoView>() {
-                            @Override
-                            public void run(@NonNull PlayVideoView view) {
-                                view.showParsingDialog();
-                            }
-                        });
+                        ifViewAttached(PlayVideoView::showParsingDialog);
                     }
 
                     @Override
                     public void onSuccess(final VideoResult videoResult) {
                         dataManager.resetPorn91VideoWatchTime(false);
-                        ifViewAttached(new ViewAction<PlayVideoView>() {
-                            @Override
-                            public void run(@NonNull PlayVideoView view) {
-                                view.parseVideoUrlSuccess(saveVideoUrl(videoResult, v9PornItem));
-                            }
-                        });
+                        ifViewAttached(view -> view.parseVideoUrlSuccess(saveVideoUrl(videoResult, v9PornItem)));
                     }
 
                     @Override
                     public void onError(final String msg, int code) {
-                        ifViewAttached(new ViewAction<PlayVideoView>() {
-                            @Override
-                            public void run(@NonNull PlayVideoView view) {
-                                view.errorParseVideoUrl(msg);
-                            }
-                        });
+                        ifViewAttached(view -> view.errorParseVideoUrl(msg));
                     }
                 });
+    }
+
+    /**
+     * 需要在UI线程执行
+     * 借助webView, 动态加载md5.js，传入相关的参数也是可用解析得到地址
+     *
+     * @param mWebView webView
+     */
+    private void decodeUrl(WebView mWebView) {
+        String a = "MXoqQlMPfiwrPSYKNCFiWwVRCldRCgZffBdgKTZzBiYiNlU/IgcMQXwuPU8CT2FbLAkTS3hVGAQoHjEQOSFzQBYCKFwOfStgHCECTmZyMhg+YXovMAwdEjw6Lw8GVzQmDBAMIjYSPAsnHQ1YJTUjLx0gTFQFCScoIQQ9RgIlD0wLf3EIbAY9BCF2d0cvcQcf";
+        String b = "a2d47W4FqndpWL/bOcbg5BGi0nXQy7SSoL2JoSA41zp8N6X/OMB14/UsfdVgtHF4uFysmNzYKtez57ZIkSKFTKKEfVuUbgXJZGdVcAfgwIHikanWSt+eKMrFhLosabZuAL+x6AkrmDF0";
+        //Javascript返回add()函数的计算结果。
+        mWebView.evaluateJavascript("parserVideoUrl('" + a + "','" + b + "')", value -> {
+            Logger.t(TAG).d(value);
+            if (TextUtils.isEmpty(value)) {
+                return;
+            }
+            Document source = Jsoup.parse(value.replace("\\u003C", "<"));
+            String videoUrl = source.select("source").first().attr("src");
+            Logger.t(TAG).d(videoUrl);
+        });
     }
 
     @Override
@@ -151,22 +158,12 @@ public class PlayVideoPresenter extends MvpBasePresenter<PlayVideoView> implemen
         downloadPresenter.downloadVideo(v9PornItem, isForceReDownload, new DownloadPresenter.DownloadListener() {
             @Override
             public void onSuccess(final String message) {
-                ifViewAttached(new ViewAction<PlayVideoView>() {
-                    @Override
-                    public void run(@NonNull PlayVideoView view) {
-                        view.showMessage(message, TastyToast.SUCCESS);
-                    }
-                });
+                ifViewAttached(view -> view.showMessage(message, TastyToast.SUCCESS));
             }
 
             @Override
             public void onError(final String message) {
-                ifViewAttached(new ViewAction<PlayVideoView>() {
-                    @Override
-                    public void run(@NonNull PlayVideoView view) {
-                        view.showMessage(message, TastyToast.ERROR);
-                    }
-                });
+                ifViewAttached(view -> view.showMessage(message, TastyToast.ERROR));
             }
         });
     }
@@ -176,22 +173,12 @@ public class PlayVideoPresenter extends MvpBasePresenter<PlayVideoView> implemen
         favoritePresenter.favorite(uId, videoId, ownnerId, new FavoritePresenter.FavoriteListener() {
             @Override
             public void onSuccess(String message) {
-                ifViewAttached(new ViewAction<PlayVideoView>() {
-                    @Override
-                    public void run(@NonNull PlayVideoView view) {
-                        view.favoriteSuccess();
-                    }
-                });
+                ifViewAttached(PlayVideoView::favoriteSuccess);
             }
 
             @Override
             public void onError(final String message) {
-                ifViewAttached(new ViewAction<PlayVideoView>() {
-                    @Override
-                    public void run(@NonNull PlayVideoView view) {
-                        view.showError(message);
-                    }
-                });
+                ifViewAttached(view -> view.showError(message));
             }
         });
     }
