@@ -1,16 +1,12 @@
 package com.u9porn.ui.porn9video.videolist;
 
 import android.arch.lifecycle.Lifecycle;
-import android.support.annotation.NonNull;
 
 import com.hannesdorfmann.mosby3.mvp.MvpBasePresenter;
 import com.orhanobut.logger.Logger;
 import com.trello.rxlifecycle2.LifecycleProvider;
 import com.u9porn.data.DataManager;
-import com.u9porn.data.model.BaseResult;
 import com.u9porn.data.db.entity.V9PornItem;
-import com.u9porn.di.PerActivity;
-import com.u9porn.di.PerFragment;
 import com.u9porn.rxjava.CallBackWrapper;
 import com.u9porn.rxjava.RetryWhenProcess;
 import com.u9porn.rxjava.RxSchedulersHelper;
@@ -22,7 +18,6 @@ import javax.inject.Inject;
 
 import io.reactivex.Observable;
 import io.reactivex.disposables.Disposable;
-import io.reactivex.functions.Function;
 
 /**
  * @author flymegoc
@@ -63,31 +58,27 @@ public class VideoListPresenter extends MvpBasePresenter<VideoListView> implemen
         if ("watch".equalsIgnoreCase(category)) {
             //最近更新
             action(dataManager.loadPorn9VideoRecentUpdates(category, page, cleanCache, isLoadMoreCleanCache)
-                    .map(new Function<BaseResult<List<V9PornItem>>, List<V9PornItem>>() {
-                        @Override
-                        public List<V9PornItem> apply(BaseResult<List<V9PornItem>> baseResult) throws Exception {
-                            if (page == 1) {
-                                totalPage = baseResult.getTotalPage();
-                            }
-                            return baseResult.getData();
+                    .map(baseResult -> {
+                        if (page == 1) {
+                            totalPage = baseResult.getTotalPage();
                         }
+                        return baseResult.getData();
                     }), pullToRefresh, skipPage);
         } else {
             //其他栏目
             if (!"top1".equals(category)) {
                 m = null;
             } else {
+                //上月最热
+                category="top";
                 m = "-1";
             }
             Observable<List<V9PornItem>> ob = dataManager.loadPorn9VideoByCategory(category, viewType, page, m, cleanCache, isLoadMoreCleanCache)
-                    .map(new Function<BaseResult<List<V9PornItem>>, List<V9PornItem>>() {
-                        @Override
-                        public List<V9PornItem> apply(BaseResult<List<V9PornItem>> baseResult) throws Exception {
-                            if (page == 1) {
-                                totalPage = baseResult.getTotalPage();
-                            }
-                            return baseResult.getData();
+                    .map(baseResult -> {
+                        if (page == 1) {
+                            totalPage = baseResult.getTotalPage();
                         }
+                        return baseResult.getData();
                     });
             action(ob, pullToRefresh, skipPage);
         }
@@ -103,53 +94,47 @@ public class VideoListPresenter extends MvpBasePresenter<VideoListView> implemen
         return dataManager.isOpenSkipPage();
     }
 
-    private void action(Observable<List<V9PornItem>> observable, final boolean pullToRefresh, final int skipPage) {
+    private void action(Observable<List<V9PornItem>> observable, boolean pullToRefresh, int skipPage) {
         observable.retryWhen(new RetryWhenProcess(RetryWhenProcess.PROCESS_TIME))
-                .compose(RxSchedulersHelper.<List<V9PornItem>>ioMainThread())
-                .compose(provider.<List<V9PornItem>>bindUntilEvent(Lifecycle.Event.ON_DESTROY))
+                .compose(RxSchedulersHelper.ioMainThread())
+                .compose(provider.bindUntilEvent(Lifecycle.Event.ON_DESTROY))
                 .subscribe(new CallBackWrapper<List<V9PornItem>>() {
                     @Override
                     public void onBegin(Disposable d) {
                         //首次加载显示加载页
-                        ifViewAttached(new ViewAction<VideoListView>() {
-                            @Override
-                            public void run(@NonNull VideoListView view) {
-                                if (page == 1 && !pullToRefresh && skipPage != 1) {
-                                    view.showLoading(pullToRefresh);
-                                }
-                                if (skipPage > 0) {
-                                    view.showSkipPageLoading();
-                                }
+                        ifViewAttached(view -> {
+                            if (page == 1 && !pullToRefresh && skipPage != 1) {
+                                view.showLoading(pullToRefresh);
+                            }
+                            if (skipPage > 0) {
+                                view.showSkipPageLoading();
                             }
                         });
                     }
 
                     @Override
                     public void onSuccess(final List<V9PornItem> v9PornItems) {
-                        ifViewAttached(new ViewAction<VideoListView>() {
-                            @Override
-                            public void run(@NonNull VideoListView view) {
-                                if (page == 1 || skipPage > 0) {
-                                    view.setData(v9PornItems);
-                                    view.showContent();
-                                    if (page == 1) {
-                                        view.setPageData(getPageList());
-                                    }
-                                    view.updateCurrentPage(page);
-                                } else {
-                                    view.loadMoreDataComplete();
-                                    view.setMoreData(v9PornItems);
-                                    view.updateCurrentPage(page);
+                        ifViewAttached(view -> {
+                            if (page == 1 || skipPage > 0) {
+                                view.setData(v9PornItems);
+                                view.showContent();
+                                if (page == 1) {
+                                    view.setPageData(getPageList());
                                 }
-                                if (skipPage > 0) {
-                                    view.hideSkipPageLoading();
-                                }
-                                //已经最后一页了
-                                if (page >= totalPage) {
-                                    view.noMoreData();
-                                } else {
-                                    page++;
-                                }
+                                view.updateCurrentPage(page);
+                            } else {
+                                view.loadMoreDataComplete();
+                                view.setMoreData(v9PornItems);
+                                view.updateCurrentPage(page);
+                            }
+                            if (skipPage > 0) {
+                                view.hideSkipPageLoading();
+                            }
+                            //已经最后一页了
+                            if (page >= totalPage) {
+                                view.noMoreData();
+                            } else {
+                                page++;
                             }
                         });
                     }
@@ -157,17 +142,14 @@ public class VideoListPresenter extends MvpBasePresenter<VideoListView> implemen
                     @Override
                     public void onError(final String msg, int code) {
                         //首次加载失败，显示重试页
-                        ifViewAttached(new ViewAction<VideoListView>() {
-                            @Override
-                            public void run(@NonNull VideoListView view) {
-                                if (page == 1) {
-                                    view.showError(msg);
-                                } else {
-                                    view.loadMoreFailed();
-                                }
-                                if (skipPage > 0) {
-                                    view.hideSkipPageLoading();
-                                }
+                        ifViewAttached(view -> {
+                            if (page == 1) {
+                                view.showError(msg);
+                            } else {
+                                view.loadMoreFailed();
+                            }
+                            if (skipPage > 0) {
+                                view.hideSkipPageLoading();
                             }
                         });
                     }
